Refactor the whole project structure
