+++ lib/common/key.dart
@@ -1,0 +1,1 @@
+const apiKey = 'c598b23ae93ac54b9a2d4ec6e2746f1e';
