+++ lib/common/network_helper.dart
@@ -1,0 +1,20 @@
+import 'dart:convert';
+import 'package:http/http.dart' as http;
+
+class NetworkHelper {
+  NetworkHelper(this.url);
+
+  final String url;
+
+  Future getData() async {
+    http.Response response = await http.get(url);
+
+    if (response.statusCode == 200) {
+      String data = response.body;
+      print(data);
+      return jsonDecode(data);
+    } else {
+      print(response.statusCode);
+    }
+  }
+}
