+++ lib/common/url.dart
@@ -1,0 +1,1 @@
+const openWeatherMapURL = 'https://api.openweathermap.org/data/2.5/weather';
