+++ lib/main.dart
@@ -1,11 +1,10 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
-import 'package:flutter_hive/screens/homePage.dart';
+import 'package:flutter_hive/views/home_page.dart';
 import 'package:hive/hive.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:http/http.dart' as http;
-import 'dart:convert';
-Location location = Location();
+
+
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   Directory dir = await getApplicationDocumentsDirectory();
@@ -23,107 +22,5 @@ class MyApp extends StatelessWidget {
       theme: ThemeData.dark(),
       home: MyHomePage(title: 'Home Page'),
     );
-  }
-}
-
-class NEWSCREEN extends StatefulWidget {
-  @override
-  _NEWSCREENState createState() => _NEWSCREENState();
-}
-
-class _NEWSCREENState extends State<NEWSCREEN> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('FLUTTER HIVE'),
-        centerTitle: true,
-      ),
-      backgroundColor: Colors.black,
-      body: Container(
-        child: Column(
-          children: [
-            Text('data')
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-//loaction
-class Location {
-  double latitude;
-  double longitude;
-
-  Future<void> getCurrentLocation() async {
-    try {
-      Position position = await Geolocator()
-          .getCurrentPosition(desiredAccuracy: LocationAccuracy.low);
-
-      latitude = position.latitude;
-      longitude = position.longitude;
-    } catch (e) {
-      print(e);
-    }
-  }
-}
-//networking
-
-
-class NetworkHelper {
-  NetworkHelper(this.url);
-
-  final String url;
-
-  Future getData() async {
-    http.Response response = await http.get(url);
-
-    if (response.statusCode == 200) {
-      String data = response.body;
-      print(data);
-      return jsonDecode(data);
-    } else {
-      print(response.statusCode);
-    }
-  }
-}
-
-//weather
-
-const apiKey = 'c598b23ae93ac54b9a2d4ec6e2746f1e';
-const openWeatherMapURL = 'https://api.openweathermap.org/data/2.5/weather';
-
-class WeatherModel {
-  
-  Future<dynamic> getLocationWeather() async {
-    Location location = Location();
-    await location.getCurrentLocation();
-
-    NetworkHelper networkHelper = NetworkHelper(
-        '$openWeatherMapURL?lat=${location.latitude}&lon=${location.longitude}&appid=$apiKey&units=metric');
-
-    var weatherData = await networkHelper.getData();
-
-    return weatherData;
-  }
-  String getWeatherIcon(int condition) {
-    if (condition < 300) {
-      return '🌩';
-    } else if (condition < 400) {
-      return '🌧';
-    } else if (condition < 600) {
-      return '☔️';
-    } else if (condition < 700) {
-      return '☃️';
-    } else if (condition < 800) {
-      return '🌫';
-    } else if (condition == 800) {
-      return '☀️';
-    } else if (condition <= 804) {
-      return '☁️';
-    } else {
-      return '🤷‍';
-    }
   }
 }
