+++ lib/models/location.dart
@@ -1,0 +1,18 @@
+import 'package:geolocator/geolocator.dart';
+
+class Location {
+  double latitude;
+  double longitude;
+
+  Future<void> getCurrentLocation() async {
+    try {
+      Position position = await Geolocator()
+          .getCurrentPosition(desiredAccuracy: LocationAccuracy.low);
+
+      latitude = position.latitude;
+      longitude = position.longitude;
+    } catch (e) {
+      print(e);
+    }
+  }
+}
