+++ lib/views/new_screen.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class NewScreen extends StatefulWidget {
+  @override
+  _NewScreenState createState() => _NewScreenState();
+}
+
+class _NewScreenState extends State<NewScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('FLUTTER HIVE'),
+        centerTitle: true,
+      ),
+      backgroundColor: Colors.black,
+      body: Container(
+        child: Column(
+          children: [
+            Text('data')
+          ],
+        ),
+      ),
+    );
+  }
+}
